Add files via upload
This is a "skeleton" for our final .Rmd file. Contains all major points and sub-points of the CRISP-DM methodology. And our names and emails.

The idea is, rather than trying to add these points while we go, and thereby missing most of them, we will start with them all listed out, and try to satisfy each before finalizing the project for submission. Some of them may be combined into broader headings before submission, if required.